gemini/requirements.txt
crewai
crewai[tools]
PyGithub
GitPython
APScheduler
python-dotenv
langchain-google-genai
